requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
Django==5.0.6
django-appconf==1.0.6
django-cleanup==8.1.0
django-compressor==4.5.1
django-environ==0.11.2
gunicorn==22.0.0
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
packaging==24.0
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pygame==2.5.2
PySocks==1.7.1
rcssmin==1.1.2
rjsmin==1.2.2
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
